chore: move tests to tests folder, update code
